refactor: flatten XGo executor frontend files
