+++ spx-gui/src/utils/xgo-executor/index.ts
@@ -1,2 +1,2 @@
-export { XGoExecutor, type XGoCapability, type XGoExecutorOptions, type XGoFramework } from './core/executor'
-export type { XGoErrorPhase, XGoExitReason } from './core/protocol'
+export { XGoExecutor, type XGoCapability, type XGoExecutorOptions, type XGoFramework } from './executor'
+export type { XGoErrorPhase, XGoExitReason } from './protocol'
