+++ spx-gui/src/utils/xgo-executor/worker.ts
